text render slang helpers

+++ Mikoto/Resources/Resources/Shaders/vulkan-spirv/material/Material.glsl
@@ -1,17 +1,20 @@
 struct ShaderMaterial {
+    // Factors
     vec4 BaseColorFactor;
     vec4 EmissiveFactor;
     vec4 DiffuseFactor;
     vec4 SpecularFactor;
 
-    float Workflow;
+    int Workflow;
 
+    // Parameters
     float MetallicFactor;
     float RoughnessFactor;
     float AlphaMask;
     float AlphaMaskCutoff;
     float EmissiveStrength;
 
+    // UV SET
     int BaseColorTextureSet;
     int PhysicalDescriptorTextureSet;
     int NormalTextureSet;
